* Fixed angular moment pairing.

+++ src/t4c_electron_repulsion/DiagonalElectronRepulsionFunc.hpp
@@ -39,7 +39,7 @@ diag_compute(      T&                  distributor,
     
     // angular order - 0
 
-    if (angmoms == std::pair<size_t, size_t>({0, 0}))
+    if (angmoms == std::pair<int, int>({0, 0}))
     {
         erirec::comp_diag_electron_repulsion_ssss(distributor, gto_pair_block, range);
         
@@ -48,7 +48,7 @@ diag_compute(      T&                  distributor,
     
     // angular order - 1
     
-    if (angmoms == std::pair<size_t, size_t>({0, 1}))
+    if (angmoms == std::pair<int, int>({0, 1}))
     {
         erirec::comp_diag_electron_repulsion_spsp(distributor, gto_pair_block, range);
         
@@ -57,14 +57,14 @@ diag_compute(      T&                  distributor,
     
     // angular order - 2
     
-    if (angmoms == std::pair<size_t, size_t>({0, 2}))
+    if (angmoms == std::pair<int, int>({0, 2}))
     {
         erirec::comp_diag_electron_repulsion_sdsd(distributor, gto_pair_block, range);
         
         return;
     }
     
-    if (angmoms == std::pair<size_t, size_t>({1, 1}))
+    if (angmoms == std::pair<int, int>({1, 1}))
     {
         erirec::comp_diag_electron_repulsion_pppp(distributor, gto_pair_block, range);
         
@@ -73,14 +73,14 @@ diag_compute(      T&                  distributor,
     
     // angular order - 3
     
-    if (angmoms == std::pair<size_t, size_t>({0, 3}))
+    if (angmoms == std::pair<int, int>({0, 3}))
     {
         erirec::comp_diag_electron_repulsion_sfsf(distributor, gto_pair_block, range);
         
         return;
     }
     
-    if (angmoms == std::pair<size_t, size_t>({1, 2}))
+    if (angmoms == std::pair<int, int>({1, 2}))
     {
         erirec::comp_diag_electron_repulsion_pdpd(distributor, gto_pair_block, range);
         
@@ -89,21 +89,21 @@ diag_compute(      T&                  distributor,
     
     // angular order - 4
     
-    if (angmoms == std::pair<size_t, size_t>({0, 4}))
+    if (angmoms == std::pair<int, int>({0, 4}))
     {
         erirec::comp_diag_electron_repulsion_sgsg(distributor, gto_pair_block, range);
         
         return;
     }
     
-    if (angmoms == std::pair<size_t, size_t>({1, 3}))
+    if (angmoms == std::pair<int, int>({1, 3}))
     {
         erirec::comp_diag_electron_repulsion_pfpf(distributor, gto_pair_block, range);
         
         return;
     }
     
-    if (angmoms == std::pair<size_t, size_t>({2, 2}))
+    if (angmoms == std::pair<int, int>({2, 2}))
     {
         erirec::comp_diag_electron_repulsion_dddd(distributor, gto_pair_block, range);
         
@@ -112,14 +112,14 @@ diag_compute(      T&                  distributor,
     
     // angular order - 5
     
-    if (angmoms == std::pair<size_t, size_t>({1, 4}))
+    if (angmoms == std::pair<int, int>({1, 4}))
     {
         erirec::comp_diag_electron_repulsion_pgpg(distributor, gto_pair_block, range);
         
         return;
     }
     
-    if (angmoms == std::pair<size_t, size_t>({2, 3}))
+    if (angmoms == std::pair<int, int>({2, 3}))
     {
         erirec::comp_diag_electron_repulsion_dfdf(distributor, gto_pair_block, range);
         
@@ -128,14 +128,14 @@ diag_compute(      T&                  distributor,
     
     // angular order - 6
     
-    if (angmoms == std::pair<size_t, size_t>({2, 4}))
+    if (angmoms == std::pair<int, int>({2, 4}))
     {
         erirec::comp_diag_electron_repulsion_dgdg(distributor, gto_pair_block, range);
         
         return;
     }
     
-    if (angmoms == std::pair<size_t, size_t>({3, 3}))
+    if (angmoms == std::pair<int, int>({3, 3}))
     {
         erirec::comp_diag_electron_repulsion_ffff(distributor, gto_pair_block, range);
         
@@ -144,7 +144,7 @@ diag_compute(      T&                  distributor,
     
     // angular order - 7
     
-    if (angmoms == std::pair<size_t, size_t>({3, 4}))
+    if (angmoms == std::pair<int, int>({3, 4}))
     {
         erirec::comp_diag_electron_repulsion_fgfg(distributor, gto_pair_block, range);
         
@@ -153,7 +153,7 @@ diag_compute(      T&                  distributor,
     
     // angular order - 8
     
-    if (angmoms == std::pair<size_t, size_t>({4, 4}))
+    if (angmoms == std::pair<int, int>({4, 4}))
     {
         erirec::comp_diag_electron_repulsion_gggg(distributor, gto_pair_block, range);
